Update the contents of the box at the top
An informatuon session titled “Information Session on ICT Systems and Tools for UTokyo Members (Conducted in English)” will be held on September 13 at 15:00. The session will be conducted in English. Materials will be provided by the day of the session, and the recording will be uploaded later. Please refer to this page as well.
↓
An information session titled “Information Session on ICT Systems and Tools for UTokyo Members (Conducted in English)” was held on September 13, 2024. Materials and recordings of the session are provided on the linked page. Please refer to this page as well.